VASP: review fixes
